+++ R/mri_info.R
@@ -9,8 +9,10 @@
 #' arguments to \code{...}
 #' @export
 #' @examplesIf have_fs()
+#' \dontrun{
 #' img = oro.nifti::nifti(array(rnorm(5*5*5), dim = c(5,5,5)))
 #' mri_info(img)
+#' }
 mri_info <- function(file, ...) {
   fs_cmd(func = "mri_info", file = file, outfile = file, ...)
 }
@@ -18,8 +20,6 @@ mri_info <- function(file, ...) {
 #' @title MRI information Help
 #' @description This calls Freesurfer's \code{mri_info} help
 #'
-#' @param display Logical; whether to display help output
-#' @param warn Logical; whether to warn if help is not available
 #' @param ... Additional arguments passed to [fs_help()]
 #' @export
 mri_info.help <- function(...) {