+++ R/mri_synthstrip.R
@@ -11,9 +11,11 @@
 #' @return Character or nifti depending on \code{retimg}
 #' @export
 #' @examplesIf have_fs()
+#' \dontrun{
 #' mock_nifti = array(rnorm(5*5*5), dim = c(5,5,5))
 #' img = oro.nifti::nifti(mock_nifti)
 #' res = mri_synthstrip(img)
+#' }
 mri_synthstrip <- function(
   file,
   outfile = NULL,