+++ R/mris_convert_curv.R
@@ -12,6 +12,7 @@
 #' @return Result of \code{\link{mris_convert}}
 #' @export
 #' @examplesIf have_fs()
+#' \dontrun{
 #'  bert_surf_dir = file.path(fs_subj_dir(), "bert", "surf")
 #'  asc_file = mris_convert_curv(
 #'  infile = file.path(bert_surf_dir, "lh.white"),
@@ -24,6 +25,7 @@
 #'  res = read_fs_table(asc_file, header = FALSE)
 #'  colnames(res) = c("index", "coord_1", "coord_2", "coord_3", "value")
 #'  head(res)
+#' }
 mris_convert_curv <- function(
   curv,
   opts = "",