+++ R/mris_convert_normals.R
@@ -7,11 +7,13 @@
 #' @return Result of \code{\link{mris_convert}}
 #' @export
 #' @examplesIf have_fs()
+#' \dontrun{
 #'  bert_dir = file.path(fs_subj_dir(), "bert")
 #'  asc_file = mris_convert_normals(
 #'  infile = file.path(bert_dir, "surf", "lh.white")
 #'  )
 #' readLines(asc_file, n = 6)
+#' }
 mris_convert_normals <- function(
   opts = "",
   ...